Code refactor and added a "autoPublishChannel" property in config.js
